requirements.txt
pymupdf
faiss-cpu
numpy
langchain
pytesseract
Pillow
rank-bm25
sentence-transformers
transformers
torch
psutil
scikit-learn
langchain-community
huggingface_hub


